+++ portal.cpp
@@ -1,7 +1,8 @@
 #include "portal.h"
 
-Portal::Portal(const std::shared_ptr<Map> &map, const int x, const int y,
-               const int new_cell_x, const int new_cell_y)
+Portal::Portal(const std::shared_ptr<Map>& map,
+               const size_t x, const size_t y,
+               const size_t new_cell_x, const size_t new_cell_y)
     : ObjectOnMap(map, x, y), new_cell_x_(new_cell_x), new_cell_y_(new_cell_y) {
   LoadImage();
 }
@@ -11,5 +12,5 @@ void Portal::LoadImage() {
   scaled_image_ = image_;
 }
 
-int Portal::GetNewCellX() const { return new_cell_x_; }
-int Portal::GetNewCellY() const { return new_cell_y_; }
+size_t Portal::GetNewCellX() const { return new_cell_x_; }
+size_t Portal::GetNewCellY() const { return new_cell_y_; }
